chore: refactor profile image upload and remove cover image
